+++ docs/changelog.rst
@@ -15,6 +15,9 @@ v1.6.0 (in development)
   - `~PyPISimple.get_provenance()` method added to `PyPISimple`
   - `NoProvenanceError` exception type added
 
+- Add ``url`` fields to the `DigestMismatchError`, `NoDigestsError`, and
+  `NoMetadataError` classes
+
 
 v1.5.0 (2024-02-24)
 -------------------